add StartNamedTunnel function and Stop() function for NamedTunnel

+++ internal/tunnel/cloudflare.go
@@ -15,6 +15,10 @@ type QuickTunnel struct {
 	PublicURL string
 }
 
+type NamedTunnel struct {
+	Cmd *exec.Cmd
+}
+
 // start a quick tunnel to localhost using cloudflare
 func StartQuickTunnel(ctx context.Context, localURL string) (*QuickTunnel, error) {
 	// find cloudflare.exe 
@@ -70,5 +74,34 @@ func (t *QuickTunnel) Stop() error {
 		return nil
 	}
 
+	return t.Cmd.Process.Kill()
+}
+
+
+// start a named tunnel using a token from Cloudflare ZeroTrust
+func StartNamedTunnel(ctx context.Context, token string) (*NamedTunnel, error) {
+	// find cloudflare.exe 
+	exePath, err := os.Executable()
+	if err != nil {
+		return nil, err
+	}
+	exeDir := filepath.Dir(exePath)
+	cloudflarePath := filepath.Join(exeDir, "cloudflared.exe")
+
+	cmd := exec.CommandContext(ctx, cloudflarePath, "tunnel" , "run", "--token", token )
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Start(); err != nil {
+		return nil, fmt.Errorf("start cloudflared: %w", err)
+	}
+	return &NamedTunnel{Cmd: cmd}, nil
+}
+
+// stop Named tunnel
+func (t *NamedTunnel) Stop() error {
+	if t == nil || t.Cmd == nil || t.Cmd.Process == nil {
+		return nil
+	}
 	return t.Cmd.Process.Kill()
 }
